requirements.txt
Flask
Pillow
piexif
gunicorn